requirements.txt
selenium
Appium-Python-Client

requests
python-box
PyYAML